requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django==4.2
django-crispy-forms==2.0
django-extensions==3.2.1
Pillow==9.5.0
pypng==0.20220715.0
qrcode==7.4.2
reportlab==3.6.12
sqlparse==0.4.4
typing_extensions==4.5.0
